Move diffcmd to run package
